return getElement from utils to component and composable.
